# sse2 for few functions in bandpass filter