- do not reallocate MDA data when source has the same size in `copy_from`
- removed temporary array for grid values in `define_kernel` in `bandpass_filter`
- vectorized some multiplication loops in `bandpass_filter` with sse2
- MDA allocates aligned memory when compiling with SSE

Verified results by comparing the output files from `bandpass_filter` before and after this update (`diff` reports no change in them). Unit tests for MDA passed too.

Example time measurements for the bandpass step:
```
before:
228,84 sec
218,969 sec
231,087 sec
230,892 sec
203,238 sec

after:
194,342 sec
201,884 sec
194,582 sec
205,13 sec
195,753 sec
```

Not much of a speed up, but this seems to be dominated by fft anyway.

@magland I worked on this code as I could not run any steps after the `whiten` processor.... I will try again tomorrow by doing a fresh install.